more warning removals for mac

+++ Source/InformationObjectDefinition/gdcmTableReader.cxx
@@ -555,8 +555,8 @@ int TableReader::Read()
   int ret = 0;
   do {
     is.read(buf, sizeof(buf));
-	std::streamsize len = is.gcount();
-    done = len < sizeof(buf);
+	  std::streamsize len = is.gcount();
+    done = (unsigned int)len < sizeof(buf);
     if (XML_Parse(parser, buf, (int)len, done) == XML_STATUS_ERROR) {
       fprintf(stderr,
         "%s at line %" XML_FMT_INT_MOD "u\n",
